pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lmtwt"
version = "0.1.0"
description = "Let Me Talk With Them — security testing framework for LLM prompt injection and jailbreak resistance"
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "Tanuphat Tin", email = "tanuphat.chai@gmail.com" },
]
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Information Technology",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Security",
]
dependencies = [
  "openai>=1.0.0",
  "anthropic>=0.5.0",
  "google-genai>=1.0",
  "httpx>=0.27",
  "python-dotenv>=1.0.0",
  "rich>=13.0.0",
  "tenacity>=8.2.0",
  "aiolimiter>=1.1",
  "pydantic>=2.0",
  "matplotlib>=3.7.0",
  "pandas>=1.5.0",
]

[project.optional-dependencies]
local = [
  "torch>=2.0.0",
  "transformers>=4.30.0",
  "accelerate>=0.20.0",
]
web = [
  "gradio>=3.50.0",
]

[dependency-groups]
dev = [
  "pytest>=7.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=4.0",
  "ruff>=0.5",
]

[tool.uv]
# Default install brings the dev group along (uv-only behavior).
# Plain `pip install -e .` still gives just the runtime deps.
default-groups = ["dev"]

[project.scripts]
lmtwt = "lmtwt.cli:main"

[project.urls]
Homepage = "https://github.com/tanuphattin/LMTWT"
Issues = "https://github.com/tanuphattin/LMTWT/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/lmtwt"]

[tool.hatch.build.targets.sdist]
include = [
  "src/lmtwt",
  "tests",
  "examples",
  "docs",
  "README.md",
  "LICENSE",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src"]

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "F",   # pyflakes
  "W",   # pycodestyle warnings
  "I",   # isort
  "UP",  # pyupgrade
]
ignore = [
  "E501",  # line-too-long (handled by formatter where it matters)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F401"]  # unused imports common in test fixtures

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
